refactor: refactored all of methods in the application

+++ src/main/java/br/com/edielsonassis/bookstore/services/BookService.java
@@ -1,20 +1,23 @@
 package br.com.edielsonassis.bookstore.services;
 
-import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.core.io.Resource;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Direction;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import br.com.edielsonassis.bookstore.dtos.v1.request.BookRequest;
 import br.com.edielsonassis.bookstore.dtos.v1.request.BookUpdateRequest;
 import br.com.edielsonassis.bookstore.dtos.v1.response.BookResponse;
+import br.com.edielsonassis.bookstore.dtos.v1.response.BookUpdateResponse;
 import br.com.edielsonassis.bookstore.mapper.Mapper;
 import br.com.edielsonassis.bookstore.models.Book;
+import br.com.edielsonassis.bookstore.models.User;
 import br.com.edielsonassis.bookstore.repositories.BookRepository;
-import br.com.edielsonassis.bookstore.services.exceptions.DataBaseException;
 import br.com.edielsonassis.bookstore.services.exceptions.ObjectNotFoundException;
+import br.com.edielsonassis.bookstore.utils.component.AuthenticatedUser;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -24,18 +27,15 @@ import lombok.extern.slf4j.Slf4j;
 public class BookService {
     
     private final BookRepository repository;
+    private final FileStorageService fileService;
 
     public BookResponse createBook(BookRequest bookRequest) {
         var book = Mapper.parseObject(bookRequest, Book.class);
+        book.setUser(currentUser());
+        book.setDownloadUrl(bookRequest.getFile().getOriginalFilename());
+        fileService.storeFile(bookRequest.getFile());
         log.info("Creating a new book: {}", bookRequest.getTitle());
         return Mapper.parseObject(repository.save(book), BookResponse.class);
-
-    }
-
-    public BookResponse findBookById(Long id) {
-        log.info("Searching for book with ID: {}", id);
-        var book = findById(id);
-        return Mapper.parseObject(book, BookResponse.class);
     }
 
     public Page<BookResponse> findBookByName(String name, Integer page, Integer size, String direction) {
@@ -52,29 +52,49 @@ public class BookService {
         return repository.findAll(pageable).map(book -> Mapper.parseObject(book, BookResponse.class));
     }
 
-    public BookResponse updateBook(BookUpdateRequest bookRequest) {
-        var book = findById(bookRequest.getBookId());
+    public BookUpdateResponse updateBook(BookUpdateRequest bookRequest) {
+        var book = findBookById(bookRequest.getBookId());
+        hasPermission(book);
+        var existingBook = book;
         book = Mapper.parseObject(bookRequest, Book.class);
+        book.setDownloadUrl(existingBook.getDownloadUrl());
+        book.setUser(existingBook.getUser());
         log.info("Updating book with name: {}", book.getTitle());
-        return Mapper.parseObject(repository.save(book), BookResponse.class);
+        return Mapper.parseObject(repository.save(book), BookUpdateResponse.class);
     }
 
     public void deleteBook(Long id) {
-        var book = findById(id);
-        try {
-            log.info("Attempting to delete book with ID: {}", book.getBookId());
-            repository.delete(book);
-            log.info("Book with name: {} successfully deleted", book.getTitle());
-        } catch (DataIntegrityViolationException e) {
-            log.error("Failed to delete book with name: {} due to referential integrity violation", book.getTitle(), e.getMessage());
-            throw new DataBaseException(e.getMessage());
+        var book = findBookById(id);
+        if (!(currentUser().equals(book.getUser()) || isAdmin(currentUser()))) {
+            throw new AccessDeniedException("You do not have permission to delete books");
         }
+        log.info("Deleting book with ID: {}", book.getBookId());
+        repository.delete(book);
     }
 
-    private Book findById(Long id) {
+    public Resource getFile(String filename) {
+        return fileService.loadFileAsResource(filename);
+    }
+
+    private Book findBookById(Long id) {
         return repository.findById(id).orElseThrow(() -> {
             log.error("Book ID not found: {}", id);
             return new ObjectNotFoundException("Book not found");
         });
+    }
+
+    private boolean hasPermission(Book book) {
+        if (!(currentUser().equals(book.getUser()) || isAdmin(currentUser()))) {
+            throw new AccessDeniedException("You do not have permission to delete books");
+        }
+        return true;
+    }
+
+    private User currentUser() {
+        return AuthenticatedUser.getCurrentUser();
+    }
+
+    private boolean isAdmin(User user) {
+        return user.getPermissions().stream().anyMatch(permission -> permission.getDescription().equals("ADMIN"));
     }
 }
